[r] Add Input- and Outputport classes
Splitted the port hierarchy and added an intermediary level between
the base class 'Port' and the most derived, templated port classes.
'Inputport' and 'Outputport' are now virtual and serve as another
level of abstraction at which to work with ports, where you may want to,
for example have a container of only input ports, iterate over them, and
set their push callbacks without knowing what types they're holding.

The most derived port types are now named InputPortOf and OutputPortOf.
Working directly at this level should offer the best performance and
compile-time safety, while the higher levels trade these two for
more and more flexibility.

+++ libs/gui/include/rsp/gui/Animation.hpp
@@ -77,7 +77,7 @@ auto Animation<T>::linear(T const& startValue, T const& endValue, float p) -> T
 template <typename T>
 auto Animation<T>::spring(T const& startValue, T const& endValue, float p) -> T
 {
-	float f = std::sinf(p * 10 * 3.141f) / (p + 1);
+	float f = std::sin(p * 10 * 3.141f) / (p + 1);
 	T difference = endValue - startValue;
 	return endValue - f * difference;
 }
